requirements.txt
asgiref==3.9.1
cffi==1.17.1
click==7.1.2
cryptography==45.0.5
Django==5.2.4
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.59.0
h11==0.9.0
idna==3.10
passlib==1.7.4
pycparser==2.22
pydantic==1.10.22
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
starlette==0.13.4
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.11.5
websockets==8.1
pytest==8.3.2
httpx
requests==2.32.4